Implement new way to load student files that makes the error messages useful in case of syntax errors
TODO: replace the name of the tmp file with the code that student submitted.
Confusing: When students submitted multiple files, they get appended to each other.
This way of appending might not be ideal. We should try moving the old load student files
into the compute points method of RSpecGrader actually.