fetch_files! is unncessary
Calling to fetch_files! is unnecessary, since the only thing that it
does is replace the url of the file with it's contents(it does a get
request and puts the body of the response), and then is replaced again
with write_to_location! with the value 'submissions/', so the previous
value is never used.